yolo-service/requirements.txt
# For the YOLO service (run: pip install -r yolo-service/requirements.txt)
ultralytics>=8.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
requests>=2.28.0
python-dotenv>=1.0.0
